+++ src/i18n_check/check/non_source_keys.py
@@ -21,6 +21,7 @@ from i18n_check.utils import (
     PATH_SEPARATOR,
     config_i18n_directory,
     config_i18n_src_file,
+    config_i18n_src_file_name,
     get_all_json_files,
     read_json_file,
 )
@@ -106,15 +107,14 @@ def non_source_keys_check(
             )
             for k in non_source_keys_dict
         )
-        is_an_or_are = "is an"
-        has_or_have = "has"
-        if len(non_source_keys_dict) > 1:
-            is_an_or_are = "are"
-            has_or_have = "have"
+
+        is_an_or_are = "is an" if len(non_source_keys_dict) == 1 else "are"
+        has_or_have = "has" if len(non_source_keys_dict) == 1 else "have"
+        file_or_files = "file" if len(non_source_keys_dict) == 1 else "files"
 
         error_message = (
             "\n"
-            + f"[red]❌ non-source-keys error: There {is_an_or_are} i18n target JSON files that {has_or_have} keys that are not in the i18n source file. Please remove or rename the following keys:"
+            + f"[red]❌ non-source-keys error: There {is_an_or_are} i18n target JSON {file_or_files} that {has_or_have} keys that are not in the {config_i18n_src_file_name} i18n source file. Please remove or rename the following keys:"
             + "\n\n"
             + non_source_keys_string
             + "[/red]"