+++ docs/source/_static/CHANGELOG.rst
@@ -11,6 +11,20 @@ See the `releases for i18n-check <https://github.com/activist-org/i18n-check/rel
 
 Emojis for the following are chosen based on `gitmoji <https://gitmoji.dev/>`_.
 
+i18n-check 1.16.1
+-----------------
+
+✅ Tests
+^^^^^^^
+
+- Tests have been updated to pass given changes to the CLI outputs.
+
+♻️ Code Refactoring
+^^^^^^^^^^^^^^^^^^^
+
+- The name of the i18n source file has been added to the output for those checks that relate to it.
+- The `aria-labels` and `alt-texts` checks now include a more verbose error message that provides more context on the issues.
+
 i18n-check 1.16.0
 -----------------